Conflicts:
	dashboards-ui-test-common-api/build.gradle
	dashboards-ui-test-common-api/src/integTest/java/oracle/sysman/emaas/platform/dashboards/tests/ui/impl/DashboardHomeUtil_1130.java
	tests/suites/emcpdf_allsrg_3n.gradle
	tests/suites/emcpdf_test_3n.gradle
	tests/suites/emcpdf_upgrade_test_3n.gradle

+++ dashboards-core/src/main/java/oracle/sysman/emaas/platform/dashboards/core/model/UserOptions.java
@@ -1,7 +1,11 @@
 package oracle.sysman.emaas.platform.dashboards.core.model;
 
+import java.math.BigInteger;
+import java.util.Date;
+
 import oracle.sysman.emaas.platform.dashboards.core.exception.DashboardException;
 import oracle.sysman.emaas.platform.dashboards.core.exception.functional.CommonFunctionalException;
+import oracle.sysman.emaas.platform.dashboards.core.util.DateUtil;
 import oracle.sysman.emaas.platform.dashboards.core.util.MessageUtils;
 import oracle.sysman.emaas.platform.dashboards.entity.EmsUserOptions;
 
@@ -11,7 +15,7 @@ import oracle.sysman.emaas.platform.dashboards.entity.EmsUserOptions;
  */
 public class UserOptions {
     private String userName;
-    private Long dashboardId;
+    private BigInteger dashboardId;
     private Long autoRefreshInterval;
     private String extendedOptions;
 
@@ -23,11 +27,11 @@ public class UserOptions {
         this.userName = userName;
     }
 
-    public Long getDashboardId() {
+    public BigInteger getDashboardId() {
         return dashboardId;
     }
 
-    public void setDashboardId(Long id) {
+    public void setDashboardId(BigInteger id) {
         this.dashboardId = id;
     }
 
@@ -60,7 +64,7 @@ public class UserOptions {
         }
         userOptions.setAutoRefreshInterval(Long.valueOf(emsUserOptions.getAutoRefreshInterval()));
         userOptions.setUserName(emsUserOptions.getUserName());
-        userOptions.setDashboardId(Long.valueOf(emsUserOptions.getDashboardId()));
+        userOptions.setDashboardId(emsUserOptions.getDashboardId());
         userOptions.setExtendedOptions(emsUserOptions.getExtendedOptions());
         return userOptions;
     }
